回复功能(二)

+++ src/main/java/com/life/demo/Controller/PublishController.java
@@ -24,7 +24,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String EditPublish(@PathVariable("id") Integer id,
+    public String EditPublish(@PathVariable("id") Long id,
                               Model model) {
         QuestionDTO question = questionService.getByID(id);
         model.addAttribute("title", question.getTitle());
@@ -45,7 +45,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model //返回错误信息
     ) {
